requirements.txt
Django>=3.1.1
djangorestframework==3.11.2
requests
pytest
responses